# specs/007-management-user-auth/contracts/management-users-auth.openapi.yaml
openapi: 3.0.3
info:
  title: Task Management System API
  version: 1.2.0
  description: Planned OpenAPI contract for ManagementUser authentication and protected endpoints.
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    CreateManagementUserRequestDto:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    LoginManagementUserRequestDto:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    ManagementUserResponseDto:
      type: object
      required:
        - id
        - username
        - email
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
    ManagementUserLoginResponseDto:
      type: object
      required:
        - accessToken
        - tokenType
        - expiresAt
        - user
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
          example: Bearer
        expiresAt:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/ManagementUserResponseDto'
    CreateManagementTaskRequestDto:
      type: object
      required:
        - title
        - description
        - dueDate
        - userId
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: integer
          enum: [0, 1, 2]
        dueDate:
          type: string
          format: date-time
        userId:
          type: string
          format: uuid
    UpdateManagementTaskRequestDto:
      type: object
      required:
        - id
        - title
        - description
        - status
        - dueDate
        - userId
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        status:
          type: integer
          enum: [0, 1, 2]
        dueDate:
          type: string
          format: date-time
        userId:
          type: string
          format: uuid
    PatchManagementTaskRequestDto:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        dueDate:
          type: string
          format: date-time
        userId:
          type: string
          format: uuid
      minProperties: 1
    UpdateManagementTaskStatusRequestDto:
      type: object
      required:
        - status
      properties:
        status:
          type: integer
          enum: [0, 1, 2]
    ManagementTaskResponseDto:
      type: object
      required:
        - id
        - title
        - description
        - status
        - dueDate
        - userId
        - isArchived
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        status:
          type: integer
          enum: [0, 1, 2]
        dueDate:
          type: string
          format: date-time
        userId:
          type: string
          format: uuid
        isArchived:
          type: boolean
        archivedAt:
          type: string
          format: date-time
          nullable: true
    PagedManagementTaskResponseDto:
      type: object
      required:
        - items
        - page
        - pageSize
        - totalCount
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ManagementTaskResponseDto'
        page:
          type: integer
        pageSize:
          type: integer
        totalCount:
          type: integer
    ManagementTaskSummaryResponseDto:
      type: object
      required:
        - statuses
      properties:
        statuses:
          type: array
          items:
            $ref: '#/components/schemas/ManagementTaskStatusCountDto'
    ManagementTaskStatusCountDto:
      type: object
      required:
        - status
        - count
      properties:
        status:
          type: integer
          enum: [0, 1, 2]
        count:
          type: integer
    ErrorResponseDto:
      type: object
      required:
        - statusCode
        - message
        - traceId
      properties:
        statusCode:
          type: integer
        message:
          type: string
        traceId:
          type: string
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ValidationErrorDto'
    ValidationErrorDto:
      type: object
      required:
        - propertyName
        - message
      properties:
        propertyName:
          type: string
        message:
          type: string
paths:
  /management-users:
    post:
      summary: Create a management user
      operationId: createManagementUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateManagementUserRequestDto'
      responses:
        '201':
          description: Management user created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementUserResponseDto'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '409':
          description: Username or email already in use
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
  /management-users/login:
    post:
      summary: Login a management user
      operationId: loginManagementUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginManagementUserRequestDto'
      responses:
        '200':
          description: Management user authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementUserLoginResponseDto'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
  /management-users/{id}:
    get:
      summary: Retrieve a management user by identifier for testing
      operationId: getManagementUserById
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Management user returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementUserResponseDto'
        '404':
          description: Management user not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
  /management-tasks:
    get:
      summary: Retrieve management tasks
      operationId: getManagementTasks
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: integer
            enum: [0, 1, 2]
        - in: query
          name: userId
          schema:
            type: string
            format: uuid
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: sortBy
          schema:
            type: string
            enum: [title, status, dueDate]
        - in: query
          name: sortDirection
          schema:
            type: string
            enum: [asc, desc]
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 1
        - in: query
          name: includeArchived
          schema:
            type: boolean
      responses:
        '200':
          description: Filtered task page returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedManagementTaskResponseDto'
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
    post:
      summary: Create a management task
      operationId: createManagementTask
      security:
        - BearerAuth: []
      parameters:
        - in: header
          name: Idempotency-Key
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateManagementTaskRequestDto'
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementTaskResponseDto'
        '400':
          description: Invalid request or referenced management user does not exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '409':
          description: Idempotency key reused with different payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
  /management-tasks/summary:
    get:
      summary: Retrieve counts of tasks grouped by status
      operationId: getManagementTaskSummary
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: includeArchived
          schema:
            type: boolean
      responses:
        '200':
          description: Summary returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementTaskSummaryResponseDto'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
  /management-tasks/overdue:
    get:
      summary: Retrieve overdue management tasks
      operationId: getOverdueManagementTasks
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: includeArchived
          schema:
            type: boolean
      responses:
        '200':
          description: Overdue tasks returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ManagementTaskResponseDto'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
  /management-tasks/due-within:
    get:
      summary: Retrieve management tasks due within the requested number of days
      operationId: getManagementTasksDueWithin
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: days
          required: true
          schema:
            type: integer
            minimum: 1
        - in: query
          name: includeArchived
          schema:
            type: boolean
      responses:
        '200':
          description: Tasks due within the requested window returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ManagementTaskResponseDto'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
  /management-tasks/{id}:
    get:
      summary: Retrieve a management task by identifier
      operationId: getManagementTaskById
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Task returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementTaskResponseDto'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
    put:
      summary: Fully update a management task
      operationId: updateManagementTask
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateManagementTaskRequestDto'
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementTaskResponseDto'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '409':
          description: Task cannot be modified in its current state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
    patch:
      summary: Partially update a management task
      operationId: patchManagementTask
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchManagementTaskRequestDto'
      responses:
        '200':
          description: Task partially updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementTaskResponseDto'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '409':
          description: Task cannot be modified in its current state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
    delete:
      summary: Archive a management task
      operationId: archiveManagementTask
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Task archived
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '409':
          description: Task cannot be deleted in its current state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
  /management-tasks/{id}/status:
    patch:
      summary: Update only the status of a management task
      operationId: updateManagementTaskStatus
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateManagementTaskStatusRequestDto'
      responses:
        '200':
          description: Task status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementTaskResponseDto'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '409':
          description: Status transition not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'